+++ sequence_check.c
@@ -7,7 +7,7 @@
 struct rngod *seq;
 
 static void sequence_init(void) {
-	const int initial[] = { 2, 7, 77 };
+	const uint64_t initial[] = { 2, 7, 77 };
 
 	seq = rngod_sequence_add(3, initial);
 }
@@ -18,13 +18,13 @@ static void sequence_shutdown(void) {
 
 #define RESULT(expected_, seq_)				\
 	do {						\
-		int res_ = seq_->rand(seq_);		\
+		uint64_t res_ = seq_->rand(seq_);		\
 		ck_assert_int_eq(expected_, res_);	\
 	} while (0)
 
 #define RANGE_RESULT(expected_, seq_, min_, max_)	\
 	do {						\
-		int res_ = seq_->range(seq_, min_, max_);	\
+		uint64_t res_ = seq_->range(seq_, min_, max_);	\
 		ck_assert_int_eq(expected_, res_);	\
 	} while (0)
 
@@ -64,7 +64,7 @@ START_TEST(test_seq_repeats_last) {
 } END_TEST
 
 START_TEST(test_seq_change_updates) {
-	int nvalues[] = { 20, 4, 9,  20, 11, 11 };
+	uint64_t nvalues[] = { 20, 4, 9,  20, 11, 11 };
 	rngod_sequence_sequence_set(seq, 6, nvalues);
 	RESULT(20, seq);
 	RESULT(4, seq);
@@ -76,7 +76,7 @@ START_TEST(test_seq_change_updates) {
 } END_TEST
 
 START_TEST(test_seq_change_resets_pos) {
-	int nvalues[] = { 20, 4, 9,  20, 11, 11 };
+	uint64_t nvalues[] = { 20, 4, 9,  20, 11, 11 };
 
 	RESULT(2, seq);
 	RESULT(7, seq);
@@ -94,13 +94,13 @@ START_TEST(test_seq_change_resets_pos) {
 
 
 START_TEST(test_seq_empty_fails) {
-	int nvalues[] = { 42 };
+	uint64_t nvalues[] = { 42 };
 	ck_assert_int_eq(-1, rngod_sequence_sequence_set(seq, 0, nvalues));
 	ck_assert_int_eq(-1, rngod_sequence_sequence_set(seq, 1, NULL));
 } END_TEST
 
 START_TEST(test_seq_single_item) {
-	int nvalues[] = { 42 };
+	uint64_t nvalues[] = { 42 };
 	ck_assert_int_eq(0, rngod_sequence_sequence_set(seq, 1, nvalues));
 	RESULT(42, seq);
 	RESULT(42, seq);
@@ -108,7 +108,7 @@ START_TEST(test_seq_single_item) {
 } END_TEST
 
 START_TEST(test_seq_range_sane) {
-	int values[] = { 1, 1, 5, 10 };
+	uint64_t values[] = { 1, 1, 5, 10 };
 	rngod_sequence_sequence_set(seq, 4, values);
 	RANGE_RESULT(1, seq, 1, 10);
 	RANGE_RESULT(1, seq, 0, 3);
@@ -117,7 +117,7 @@ START_TEST(test_seq_range_sane) {
 } END_TEST
 
 START_TEST(test_seq_range_outside) {
-	int values[] = { 10 };
+	uint64_t values[] = { 10 };
 	rngod_sequence_sequence_set(seq, 1, values);
 	RANGE_RESULT(1, seq, 1, 5);
 	RANGE_RESULT(3, seq, 1, 4);
@@ -127,7 +127,7 @@ START_TEST(test_seq_range_outside) {
 } END_TEST
 
 START_TEST(test_seq_range_all) {
-	int values[] = { 0, 1, 2, 3, 4, 5, 6, 7 };
+	uint64_t values[] = { 0, 1, 2, 3, 4, 5, 6, 7 };
 	rngod_sequence_sequence_set(seq, 8, values);
 	RANGE_RESULT(0, seq, 0, 3);
 	RANGE_RESULT(1, seq, 0, 3);
